[feat] UserService 인터페이스 및 기본 구현체 작성 #20
