+++ src/main/kotlin/dev/codeswamp/core/user/domain/model/Role.kt
@@ -1,0 +1,7 @@
+package dev.codeswamp.core.user.domain.model
+
+enum class Role {
+    GUEST,
+    USER,
+    ADMIN
+}
