+++ src/main/kotlin/dev/codeswamp/core/user/domain/model/User.kt
@@ -1,10 +1,10 @@
 package dev.codeswamp.core.user.domain.model
 
-import dev.codeswamp.core.user.infrastructure.entity.Role
+import javax.management.relation.Role
 
 data class User (
     val id: Long? = null,
-    val username: String,
+    val username: String,//변경 불가(사용자 ID)
     val email: String,
     val nickname: String,
     val profileUrl: String,
